requirements.txt
pandas>=2.0.0
numpy>=1.24.0
tushare>=1.2.89
akshare>=1.11.0
baostock>=0.8.8
pyyaml>=6.0
loguru>=0.7.0
python-dateutil>=2.8.2

